% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{format_pseudo_log_tick}
\alias{format_pseudo_log_tick}
\title{Format pseudo-log axis ticks}
\usage{
format_pseudo_log_tick(x)
}
\arguments{
\item{x}{Numeric tick values.}
}
\value{
A character vector.
}
\description{
Format pseudo-log axis ticks
}
