% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bms_brand_mvp.R
\name{scale_colour_bms_engine}
\alias{scale_colour_bms_engine}
\title{BMS engine colour scale}
\usage{
scale_colour_bms_engine(..., engines = c("Sage", "GNU"))
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:scale_colour_manual]{ggplot2::scale_colour_manual()}}.}

\item{engines}{Engine names to include in the scale.}
}
\value{
A ggplot2 discrete colour scale.
}
\description{
BMS engine colour scale
}
