% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_brand_mvp.R
\name{scale_colour_bs_engine}
\alias{scale_colour_bs_engine}
\title{BS engine colour scale}
\usage{
scale_colour_bs_engine(..., engines = c("Sage", "GNU"))
}
\arguments{
\item{...}{Passed to \code{ggplot2::scale_colour_manual()}.}

\item{engines}{Engine names to include in the scale.}
}
\value{
A ggplot2 discrete colour scale.
}
\description{
BS engine colour scale
}
